more work getting js things building; add await helper function
