+++ wgpu/primitive_js.go
@@ -2,6 +2,8 @@
 
 package wgpu
 
+import "syscall/js"
+
 // mapSlice can be used to transform one slice into another by providing a
 // function to do the mapping.
 func mapSlice[S, T any](slice []S, fn func(S) T) []T {
@@ -13,6 +15,16 @@ func mapSlice[S, T any](slice []S, fn func(S) T) []T {
 		result[i] = fn(v)
 	}
 	return result
+}
+
+// await is a helper function roughly equivalent to await in JS.
+func await(promise js.Value) js.Value {
+	result := make(chan js.Value)
+	promise.Call("then", js.FuncOf(func(this js.Value, args []js.Value) any {
+		result <- args[0]
+		return nil
+	}))
+	return <-result
 }
 
 // ToJS converts this type to one that can be passed as an argument
